+++ NEWS.md
@@ -2,8 +2,8 @@
 
 ## Minor changes
 - `geom_half_point` now supports passing of positions other than `PositionJitter` and `PositionIdentity` to `transformation` - given they have a `compute_panel` function.
-- added `range_scale` parameter to `geom_half_point`. When no `width` argument is passed in `transformation_params`, `range_scale` defaults to `8`, which leads to the points using half of their allotted space. `4` would have them end up using all of their allotted space. 
-- added `nudge` parameter to `geom_half_boxplot` and `geom_half_violin`. `nudge` takes as value a non-negative scalar and adds space between the middle of the space allotted to a factor on the x-axis and the respective geom.
+- added `range_scale` parameter to `geom_half_point`. When no `width` argument is passed in `transformation_params`, `range_scale` defaults to `0.5`, which leads to the points using half of their allotted space. `1` would have them end up using all of their allotted space. 
+
 
 ## Bug Fixes
 - removed range scaling of `geom_half_point` after points were already jittered